feat: refactor search and embedding spaces to be fully modular/extendable
